requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
streamlit==1.29.0
pdfplumber==0.10.3
langchain==0.0.352
langchain-openai==0.0.2
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1 